python RAG/requirements.txt
crewai
google-generativeai
python-dotenv
langchain
langchain-community
faiss-cpu
sentence-transformers
pypdf
tqdm
numpy
